feat: Update buttplug4j to the v4 spec
This will only work with the Interface Central v3 release, currently in beta.

+++ buttplug4j/src/main/java/io/github/blackspherefollower/buttplug4j/protocol/ButtplugDeviceMessage.java
@@ -7,7 +7,7 @@ public abstract class ButtplugDeviceMessage extends ButtplugMessage {
     @JsonProperty(value = "DeviceIndex", required = true)
     private long deviceIndex;
 
-    public ButtplugDeviceMessage(final long id, final long deviceIndex) {
+    public ButtplugDeviceMessage(final int id, final long deviceIndex) {
         super(id);
         this.setDeviceIndex(deviceIndex);
     }
